refactor: remove duplicated code in WellLabelLayer stories
